Duplicate code consolidation
