+++ libs/commons-types/src/main/kotlin/gov/cdc/ocio/types/utils/DateUtils.kt
@@ -1,10 +1,13 @@
-package gov.cdc.ocio.processingstatusapi.utils
+package gov.cdc.ocio.types.utils
 
-import gov.cdc.ocio.processingstatusapi.exceptions.BadRequestException
 import org.apache.commons.lang3.time.FastDateFormat
 import java.text.ParseException
 import java.util.*
 
+
+/**
+ * A utility class for handling date operations.
+ */
 object DateUtils {
 
     private const val DATE_FORMAT = "yyyyMMdd'T'HHmmss'Z'"
@@ -17,18 +20,17 @@ object DateUtils {
      * @param dateStr String
      * @param fieldName String
      * @return Long - Epoch time in milliseconds
-     * @throws BadRequestException
+     * @throws IllegalArgumentException
      */
-    @Throws(BadRequestException::class)
+    @Throws(IllegalArgumentException::class)
     fun getEpochFromDateString(dateStr: String, fieldName: String): Long {
         try {
             // Parse the date string to a Date object
             val date = dateFormat.parse(dateStr)
             return date.time
         } catch (e: ParseException) {
-            throw BadRequestException("Failed to parse $fieldName: $dateStr.  Format should be: ${DATE_FORMAT}.")
+            throw IllegalArgumentException("Failed to parse $fieldName: $dateStr. Format should be: ${DATE_FORMAT}.")
         }
     }
-
 
 }
