+++ pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/loaders/ReportCountsLoader.kt
@@ -3,10 +3,10 @@ package gov.cdc.ocio.processingstatusapi.loaders
 import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
 import gov.cdc.ocio.processingstatusapi.models.ReportCounts
 import gov.cdc.ocio.database.models.dao.ReportDao
+import gov.cdc.ocio.database.utils.SqlClauseBuilder
 import gov.cdc.ocio.processingstatusapi.models.query.PageSummary
 import gov.cdc.ocio.processingstatusapi.models.reports.*
 import gov.cdc.ocio.processingstatusapi.utils.PageUtils
-import gov.cdc.ocio.processingstatusapi.utils.SqlClauseBuilder
 import mu.KotlinLogging
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
@@ -113,7 +113,7 @@ class ReportCountsLoader: KoinComponent {
 
         val pageSizeAsInt = pageUtils.getPageSize(pageSize)
 
-        val timeRangeWhereClause = SqlClauseBuilder().buildSqlClauseForDateRange(
+        val timeRangeWhereClause = SqlClauseBuilder.buildSqlClauseForDateRange(
             daysInterval,
             dateStart,
             dateEnd,
@@ -268,7 +268,7 @@ class ReportCountsLoader: KoinComponent {
     ): ProcessingCounts {
 
         val timeRangeWhereClause =
-            SqlClauseBuilder().buildSqlClauseForDateRange(daysInterval, dateStart, dateEnd, cPrefix)
+            SqlClauseBuilder.buildSqlClauseForDateRange(daysInterval, dateStart, dateEnd, cPrefix)
 
         // Get number completed uploading
         val numCompletedUploadingSqlQuery = (
@@ -348,7 +348,7 @@ class ReportCountsLoader: KoinComponent {
     ): List<StageCounts> {
 
         val timeRangeWhereClause =
-            SqlClauseBuilder().buildSqlClauseForDateRange(daysInterval, dateStart, dateEnd, cPrefix)
+            SqlClauseBuilder.buildSqlClauseForDateRange(daysInterval, dateStart, dateEnd, cPrefix)
 
         val rollupCountsQuery = (
                 "select "
